+++ packages/cli/uni-builder/src/shared/rsc/plugins/rsc-server-plugin.ts
@@ -1,10 +1,5 @@
 import type Webpack from 'webpack';
-import {
-  type Compilation,
-  Module,
-  type ModuleGraph,
-  NormalModule,
-} from 'webpack';
+import { type Compilation, NormalModule } from 'webpack';
 import {
   type ServerManifest,
   type ServerReferencesMap,