+++ packages/runtime/plugin-runtime/src/core/plugin/types.ts
@@ -3,6 +3,7 @@ import type {
   RuntimePluginExtends,
 } from '@modern-js/plugin-v2';
 import type { Hooks } from '@modern-js/plugin-v2/runtime';
+import type { AppConfig } from '../../common';
 import type { RuntimeContext } from '../context/runtime';
 import type { Plugin } from './base';
 
@@ -14,6 +15,6 @@ export type RuntimeExtends = Required<
 
 export type RuntimePluginFuture<Extends extends RuntimePluginExtends = {}> =
   BaseRuntimePlugin<RuntimeExtends & Extends>;
-export interface RuntimeConfig {
+export interface RuntimeConfig extends AppConfig {
   plugins?: (Plugin | RuntimePluginFuture)[];
 }