+++ packages/server/plugin-koa/tests/helpers.ts
@@ -1,39 +1,49 @@
 import path from 'path';
 import { API_DIR, ApiRouter } from '@modern-js/bff-core';
+import type { Plugin } from '@modern-js/plugin-v2';
+import { server } from '@modern-js/plugin-v2/server';
 import {
-  PluginManager,
   type ServerConfig,
   type ServerPlugin,
-  createContext,
+  type ServerPluginLegacy,
+  compatPlugin,
+  handleSetupResult,
 } from '@modern-js/server-core';
+import { assign } from '@modern-js/utils/lodash';
 
-export function createPluginManager({
+export async function serverInit({
+  plugins,
   serverConfig,
 }: {
+  plugins?: (ServerPlugin | ServerPluginLegacy)[];
   serverConfig?: ServerConfig;
-} = {}) {
-  const appContext = createContext<any>({});
-
-  const pluginManager = new PluginManager({
-    cliConfig: {
-      dev: {},
-      output: {},
-      source: {},
-      tools: {},
-      server: {},
-      html: {},
-      runtime: {},
-      bff: {},
-      security: {},
-    },
-    serverConfig,
-    appContext,
+}) {
+  const { serverContext } = await server.run({
+    plugins: [compatPlugin(), ...(plugins || [])] as Plugin[],
+    options: { appContext: {}, pwd: process.cwd() },
+    config: assign(
+      {},
+      {
+        dev: {},
+        output: {},
+        source: {},
+        tools: {},
+        server: {},
+        html: {},
+        runtime: {},
+        bff: {},
+        security: {},
+      },
+      serverConfig,
+    ),
+    handleSetupResult,
   });
+  const hooks = serverContext.pluginAPI?.getHooks();
 
-  return pluginManager;
+  return hooks as any;
 }
 
-export const APIPlugin: ServerPlugin = {
+export const APIPlugin: ServerPluginLegacy = {
   name: 'api-plugin',
 
   setup(api) {