+++ packages/solutions/module-tools/src/builder/feature/json.ts
@@ -1,5 +1,5 @@
 import { readFileSync } from 'fs';
-import type { ICompiler } from '../../types';
+import type { ICompiler, LoadResult } from '../../types';
 
 const isJsonExt = (path: string) => {
   return path.endsWith('.json');
@@ -12,7 +12,9 @@ export const json = {
     compiler.hooks.load.tapPromise({ name }, async args => {
       if (isJsonExt(args.path)) {
         return {
-          contents: readFileSync(args.path),
+          contents: readFileSync(
+            args.path,
+          ) as unknown as LoadResult['contents'],
           loader: 'copy',
         };
       }