+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @modern-js/plugin-bff
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [fb6bf9e]
+  - @modern-js/server-core@2.65.5
+  - @modern-js/server-utils@2.65.5
+  - @modern-js/create-request@2.65.5
+  - @modern-js/bff-core@2.65.5
+  - @modern-js/utils@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [0d47cb8]
+- Updated dependencies [f1cd095]
+  - @modern-js/utils@2.65.4
+  - @modern-js/server-core@2.65.4
+  - @modern-js/bff-core@2.65.4
+  - @modern-js/create-request@2.65.4
+  - @modern-js/server-utils@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes