+++ packages/document/main-doc/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/main-doc
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/sandpack-react@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- @modern-js/sandpack-react@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes