+++ packages/generator/generator-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/generator-common
 
+## 3.7.23
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.65.5
+
+## 3.7.22
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.65.4
+
 ## 3.7.21
 
 ### Patch Changes