+++ packages/generator/generator-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/generator-utils
 
+## 3.7.23
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.65.5
+- @modern-js/generator-common@3.7.23
+
+## 3.7.22
+
+### Patch Changes
+
+- @modern-js/plugin-i18n@2.65.4
+- @modern-js/generator-common@3.7.22
+
 ## 3.7.21
 
 ### Patch Changes