+++ packages/generator/generators/tailwindcss-generator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/tailwindcss-generator
 
+## 3.7.23
+
+## 3.7.22
+
 ## 3.7.21
 
 ## 3.7.20