+++ packages/generator/new-action/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/new-action
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/generator-common@3.7.23
+- @modern-js/generator-utils@3.7.23
+
+## 2.65.4
+
+### Patch Changes
+
+- @modern-js/generator-common@3.7.22
+- @modern-js/generator-utils@3.7.22
+
 ## 2.65.3
 
 ### Patch Changes