+++ packages/module/plugin-module-node-polyfill/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/plugin-module-node-polyfill
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/module-tools@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- @modern-js/module-tools@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes