+++ packages/server/plugin-worker/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @modern-js/plugin-worker
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/prod-server@2.65.5
+- @modern-js/server-utils@2.65.5
+- @modern-js/utils@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [0d47cb8]
+  - @modern-js/utils@2.65.4
+  - @modern-js/prod-server@2.65.4
+  - @modern-js/server-utils@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes