+++ packages/storybook/framework/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @modern-js/storybook
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/storybook-builder@2.65.5
+- @modern-js/utils@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [0d47cb8]
+  - @modern-js/utils@2.65.4
+  - @modern-js/storybook-builder@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes