+++ packages/toolkit/e2e/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @modern-js/e2e
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/utils@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [0d47cb8]
+  - @modern-js/utils@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes