+++ scripts/update-codesmith/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scripts/update-codesmith
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/generator-utils@3.7.23
+
+## 2.65.4
+
+### Patch Changes
+
+- @modern-js/generator-utils@3.7.22
+
 ## 2.65.3
 
 ### Patch Changes