+++ scripts/vitest-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @scripts/vitest-config
 
+## 2.65.5
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- @modern-js/tsconfig@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes