+++ tests/e2e/garfish/fixtures/main-router-v6/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @e2e/garfish-main-router-v6
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [9c30e63]
+- Updated dependencies [28b95e3]
+- Updated dependencies [58a1afd]
+  - @modern-js/plugin-garfish@2.65.5
+  - @modern-js/runtime@2.65.5
+  - @modern-js/plugin-router-v5@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [7bce153]
+  - @modern-js/runtime@2.65.4
+  - @modern-js/plugin-garfish@2.65.4
+  - @modern-js/plugin-router-v5@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes