+++ tests/integration/bff-corss-project/bff-client-app/CHANGELOG.md
@@ -1,5 +1,25 @@
 # bff-client-app
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [58a1afd]
+  - @modern-js/runtime@2.65.5
+  - @modern-js/plugin-bff@2.65.5
+  - @modern-js/plugin-express@2.65.5
+  - bff-api-app@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [7bce153]
+  - @modern-js/runtime@2.65.4
+  - @modern-js/plugin-bff@2.65.4
+  - bff-api-app@2.65.4
+  - @modern-js/plugin-express@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes