+++ tests/integration/module/fixtures/preset/function/CHANGELOG.md
@@ -1,5 +1,9 @@
 # build-preset-function-test
 
+## 2.65.5
+
+## 2.65.4
+
 ## 2.65.3
 
 ## 2.65.2