+++ tests/integration/rsbuild-hook/CHANGELOG.md
@@ -1,5 +1,23 @@
 # app-rsbuild-hooks
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [e8f7355]
+- Updated dependencies [58a1afd]
+  - @modern-js/app-tools@2.65.5
+  - @modern-js/runtime@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [7bce153]
+- Updated dependencies [93e12ae]
+  - @modern-js/runtime@2.65.4
+  - @modern-js/app-tools@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes