+++ tests/integration/source-code-build/app/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @source-code-build/app
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [58a1afd]
+  - @modern-js/runtime@2.65.5
+  - @source-code-build/components@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [7bce153]
+  - @modern-js/runtime@2.65.4
+  - @source-code-build/components@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes