+++ tests/integration/ssg/fixtures/web-server/CHANGELOG.md
@@ -1,5 +1,27 @@
 # ssg-fixtures-web-server
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [e8f7355]
+- Updated dependencies [58a1afd]
+  - @modern-js/app-tools@2.65.5
+  - @modern-js/runtime@2.65.5
+  - @modern-js/plugin-server@2.65.5
+  - @modern-js/plugin-ssg@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [7bce153]
+- Updated dependencies [93e12ae]
+  - @modern-js/runtime@2.65.4
+  - @modern-js/app-tools@2.65.4
+  - @modern-js/plugin-ssg@2.65.4
+  - @modern-js/plugin-server@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes