+++ tests/integration/temp-dir/CHANGELOG.md
@@ -1,5 +1,19 @@
 # tmp-dir
 
+## 2.65.5
+
+### Patch Changes
+
+- Updated dependencies [58a1afd]
+  - @modern-js/runtime@2.65.5
+
+## 2.65.4
+
+### Patch Changes
+
+- Updated dependencies [7bce153]
+  - @modern-js/runtime@2.65.4
+
 ## 2.65.3
 
 ### Patch Changes