+++ tests/integration/transform-lodash-import/CHANGELOG.md
@@ -1,5 +1,9 @@
 # transform-import-type-lodash
 
+## 2.65.5
+
+## 2.65.4
+
 ## 2.65.3
 
 ## 2.65.2